🚀 Initial commit
The first commit for Dommit (tentative name), a DOM implementation with the goal of being fully compliant with the WATWG DOM Living Standard. It strives to be compliant not only with the spec, but also the implementation (own properties, prototype chain structure, etc.), with Google Chrome serving as the primary reference for this.